bufHash fixed to not give negative indices

+++ Stage6/bufHash.C
@@ -13,9 +13,9 @@
 int BufHashTbl::hash(const File* file, const int pageNo)
 {
   long tmp, value;
-  tmp = (int)(long)file;  // cast of pointer to the file object to an integer
+  tmp = (long)file;  // cast of pointer to the file object to an integer
   value = (tmp + pageNo) % HTSIZE;
-  return value;
+  return (int)value;
 }
